+++ R/fcast.R
@@ -39,16 +39,12 @@ check_formula = function(formula, varnames, valnames) {
 
 deparse_formula = function(expr, varnames, allvars) {
   lvars = lapply(expr, function(this) {
-    if (is.call(this)) {
-      if (this[[1L]] == quote(`+`))
-        unlist(deparse_formula(as.list(this)[-1L], varnames, allvars))
-      else this
-    } else if (is.name(this)) {
-      if (this == quote(`...`)) {
-        subvars = setdiff(varnames, allvars)
-        lapply(subvars, as.name)
-      } else this
-    }
+    if (this %iscall% '+') {
+      unlist(deparse_formula(as.list(this)[-1L], varnames, allvars))
+    } else if (is.name(this) && this==quote(`...`)) {
+      subvars = setdiff(varnames, allvars)
+      lapply(subvars, as.name)
+    } else this
   })
   lvars = lapply(lvars, function(x) if (length(x) && !is.list(x)) list(x) else x)
 }
@@ -65,11 +61,11 @@ value_vars = function(value.var, varnames) {
 }
 
 aggregate_funs = function(funs, vals, sep="_", ...) {
-  if (is.call(funs) && funs[[1L]] == "eval")
+  if (funs %iscall% 'eval')
     funs = eval(funs[[2L]], parent.frame(2L), parent.frame(2L))
-  if (is.call(funs) && as.character(funs[[1L]]) %chin% c("c", "list")) {
+  if (funs %iscall% c('c', 'list')) {
     funs = lapply(as.list(funs)[-1L], function(x) {
-      if (is.call(x) && as.character(x[[1L]]) %chin% c("c", "list")) as.list(x)[-1L] else x
+      if (x %iscall% c('c', 'list')) as.list(x)[-1L] else x
     })
   } else funs = eval(funs, parent.frame(2L), parent.frame(2L))
   if(is.function(funs)) funs = list(funs) # needed for cases as shown in test#1700.1